# Optimize the geopotential computations by removing Pow and trig functions