New benchmark:
```
Run on (4 X 3310 MHz CPU s)
09/29/18 00:24:58
--------------------------------------------------------------------
Benchmark                             Time           CPU Iterations
--------------------------------------------------------------------
BM_ComputeGeopotentialCpp/2      241652 ns     242489 ns      11966
BM_ComputeGeopotentialCpp/3      310604 ns     309429 ns       8974
BM_ComputeGeopotentialCpp/5      616819 ns     615382 ns       4487
BM_ComputeGeopotentialCpp/10    3579013 ns    3580023 ns        780
BM_ComputeGeopotentialF90/2       72471 ns      72365 ns      39019
BM_ComputeGeopotentialF90/3      111243 ns     111694 ns      25280
BM_ComputeGeopotentialF90/5      200372 ns     201117 ns      13807
BM_ComputeGeopotentialF90/10     690917 ns     692233 ns       4079
```
which fits `601929 - 285083 n + 58270 n²`.